// models/sales_saddle_stitch.model.ts
export type SalesSaddleStitchState = {
  data: {
    sals_sadd_id: string;
    sals_sadd_date: string;
    sals_sadd_company_name: string;
    sals_sadd_tel: string;
    sals_sadd_email: string;
    sals_sadd_doc_type: string;
    sals_sadd_printing_type: string;
    sals_sadd_amount: string;
    sals_sadd_quotation_request: string;
    sals_sadd_finished_size: string;
    sals_sadd_cover: string;
    sals_sadd_text: string;
    sals_sadd_cover_paper: string;
    sals_sadd_text_paper: string;
    sals_sadd_printing: string;
    sals_sadd_cover_coating: string;
    sals_sadd_text_coating: string;
    sals_sadd_printing_volume: string;
  };
};